requirements.txt
antlr4-python3-runtime==4.8
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
black==21.10b0
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1607575130224/work
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1625814692085/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262147379/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616767007030/work
cycler==0.10.0
fonttools==4.25.0
hydra-core==1.1.1
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
isort==5.10.1
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1628003481308/work
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
numexpr @ file:///tmp/build/80754af9/numexpr_1618856529730/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264704411/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
omegaconf==2.1.1
pandas==1.3.2
pathspec==0.9.0
Pillow @ file:///tmp/build/80754af9/pillow_1625655408865/work
platformdirs==2.4.0
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==6.0
regex==2021.11.10
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
-e git+ssh://git@github.com/kimhc6028/tong_simulator.git@9cca520132b7861c4390826c56698e29630d25e3#egg=rlkit&subdirectory=rlkit
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
scipy @ file:///tmp/build/80754af9/scipy_1630606796912/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==1.2.2
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
tqdm @ file:///tmp/build/80754af9/tqdm_1629302309755/work
typing-extensions==3.10.0.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
